common error handler for socket resources
