--- src/net/UdpServer.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Esockets\net;
-
-use Esockets\base\AbstractServer;
-
-class UdpServer extends AbstractServer
-{
-
-}
